add ajax for delete answer

+++ app/controllers/answers_controller.rb
@@ -12,12 +12,7 @@ class AnswersController < ApplicationController
   end
 
   def destroy
-    if current_user.author_of?(answer)
-      answer.destroy
-      redirect_to @answer.question, notice: 'Answer successfully deleted.'
-    else
-      redirect_to @answer.question, notice: 'Only author can delete answer.'
-    end
+    answer.destroy if current_user.author_of?(answer)
   end
 
   def update
